Implement stride scheduling algorithm
This commit replaces the previous simple round-robin scheduler with the stride scheduling algo
to enable priority-based process scheduling. The new scheduler ensures that processes receive CPU time
proportional to their assigned priority, leading to a more fair and controllable execution environment.

Key changes include modifications to the process structure in proc.h, where stride and priority
fields have been added to struct proc. The main scheduler logic in proc.c is updated to select the process
with the minimum stride for execution and then increments its stride based on its priority.

New processes are initialized with a default priority of 16 and a stride of 0 in allocproc().
Child processes created via fork() now inherit the priority and stride from their parent.

Additionally, a new system call, sys_set_priority, is introduced in syscall.c. This allows a
process to set its own priority, provided it is 2 or greater.

Signed-off-by: 0wnerDied <z1281552865@gmail.com>